# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v1
components:
  schemas:
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        total:
          type: number
          format: double
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        userName:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        name:
          type: string
        company:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        stock:
          type: integer
          format: int32
        isAccepted:
          type: boolean
        date:
          type: string
          format: date-time
        views_count:
          type: integer
          format: int32
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        imageBase64:
          type: string
    Review:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date-time
    ReviewDTO:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date
    UserCreationDTO:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
        imageBase64:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
...
    CompanyStadsDTO:
      type: object
      properties:
        key:
          type: string
        value:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
servers:
  - url: https://localhost:8443
    description: Generated server url
paths:
...
  /api/v1/users:
    get:
      tags:
      - api-user-controller
      summary: Get all users
      description: Retrieve a list of all users in the system.
      operationId: getAllUsers
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - api-user-controller
      summary: Update user profile
      description: Update the details of an existing user.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - api-user-controller
      summary: Create a new user
      description: Register a new user in the system.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Username already exists
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
  /api/v1/products/{id}:
    get:
      tags:
      - api-product-controller
      summary: Get product by ID
      description: Retrieve a product by its unique ID.
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - api-product-controller
      summary: Update a product
      description: Update the details of an existing product.
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - api-product-controller
      summary: Delete a product
      description: Delete a product by its unique ID.
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/users/{id}:
    get:
      tags:
      - api-user-controller
      summary: Get user by ID
      description: Retrieve a user by their unique ID.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - api-user-controller
      summary: Delete user
      description: Delete a user by their unique ID.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/graphs/userGraph:
    get:
      tags:
      - api-graphs-controller
      summary: Get user historical orders graph
      description: Retrieve the user's historical orders data for graph display.
      operationId: displayUserGraph
      responses:
        "200":
          description: User historical orders retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: number
                  format: double
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: number
                  format: double
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: number
                  format: double
  /api/v1/graphs/companyGraph:
    get:
      tags:
      - api-graphs-controller
      summary: Get company statistics graph
      description: Retrieve the company statistics data for graph display.
      operationId: displayGraph
      responses:
        "200":
          description: Company statistics retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyStadsDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyStadsDTO"
        "404":
          description: Company data not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyStadsDTO"
...